Unsigned numeric fetch  fix
Former-commit-id: d36974675252b294234b357d00263521fc0065eb

+++ trunk/plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/model/impl/jdbc/data/JDBCNumberValueHandler.java
@@ -108,9 +108,11 @@ public class JDBCNumberValueHandler extends JDBCAbstractValueHandler implements 
                 }
                 break;
             case java.sql.Types.SMALLINT:
-                value = resultSet.getShort(index);
+                // Read int in case of unsigned shorts
+                value = resultSet.getInt(index);
                 break;
             case java.sql.Types.TINYINT:
+                // Read short in case of unsigned byte
                 value = resultSet.getShort(index);
                 break;
             case java.sql.Types.BIT:
